Reformat everything

+++ api/src/Containers.js
@@ -4,15 +4,13 @@ const path = require("path");
 const config = require("../config.json");
 
 class Container extends EventEmitter {
-
     /**
-     * @param {import("./Servers")} servers 
-     * @param {string} name 
-     * @param {string} path 
-     * @param {number} port 
+     * @param {import("./Servers")} servers
+     * @param {string} name
+     * @param {string} path
+     * @param {number} port
      */
     constructor(servers, name, path, port) {
-
         super();
 
         this.servers = servers;
@@ -30,30 +28,28 @@ class Container extends EventEmitter {
     }
 
     /**
-     * @param {"stopped"|"starting"|"started"|"stopping"} state 
+     * @param {"stopped"|"starting"|"started"|"stopping"} state
      */
     _setState(state) {
         this.state = state;
         this.emit(state);
         this.servers.saveState();
-        this.servers.gateway.clients.filter((client) => client.infos.logged).forEach((client) => client.emitEvent("SERVER_STATE", { name: this.name, state: state }));
+        this.servers.gateway.clients
+            .filter((client) => client.infos.logged)
+            .forEach((client) => client.emitEvent("SERVER_STATE", { name: this.name, state: state }));
     }
 
     /**
-     * @param {object} event 
+     * @param {object} event
      */
     async _dockerEventHandler(event) {
-
         if (event.Type !== "container" || event.Actor.Attributes.name !== this.name) return;
 
-        if (event.Action === "start")
-            this._postStart();
-        else if (event.Action === "die")
-            this._postStop();
+        if (event.Action === "start") this._postStart();
+        else if (event.Action === "die") this._postStop();
     }
 
     async _preStart() {
-
         if (this.state !== "stopped") throw new Error("Not stopped");
         this._setState("starting");
 
@@ -71,12 +67,11 @@ class Container extends EventEmitter {
     }
 
     /**
-     * @param {number} memory 
-     * @param {string} file 
-     * @param {boolean} exposed 
+     * @param {number} memory
+     * @param {string} file
+     * @param {boolean} exposed
      */
     async start(memory, file, exposed = false) {
-
         await this._preStart();
 
         this.servers.dockerEvents.on("rawEvent", this._bindDockerEventHandler);
@@ -90,7 +85,9 @@ class Container extends EventEmitter {
             Image: "openjdk:8",
             WorkingDir: "/home/server",
             HostConfig: {
-                PortBindings: { "25565/tcp": [{ HostIp: exposed ? "" : "172.17.0.1", HostPort: this.port.toString() }] },
+                PortBindings: {
+                    "25565/tcp": [{ HostIp: exposed ? "" : "172.17.0.1", HostPort: this.port.toString() }]
+                },
                 Binds: [this.path + ":/home/server"],
                 AutoRemove: true
             }
@@ -100,7 +97,6 @@ class Container extends EventEmitter {
     }
 
     _postStart() {
-
         if (this.state !== "starting") throw new Error("Not starting");
 
         this._setState("started");
@@ -108,10 +104,9 @@ class Container extends EventEmitter {
     }
 
     /**
-     * @param {string} command 
+     * @param {string} command
      */
     async stop(command) {
-
         if (this.state !== "started") throw new Error("Not started");
         this._setState("stopping");
 
@@ -121,14 +116,15 @@ class Container extends EventEmitter {
 
         const killTimeout = setTimeout(() => this.kill(), 30 * 1000);
 
-        await new Promise((resolve) => this.once("stopped", () => {
-            clearTimeout(killTimeout);
-            resolve();
-        }));
+        await new Promise((resolve) =>
+            this.once("stopped", () => {
+                clearTimeout(killTimeout);
+                resolve();
+            })
+        );
     }
 
     _postStop() {
-
         if (this.state === "stopped") throw new Error("Already stopped");
 
         this.servers.dockerEvents.removeListener("rawEvent", this._bindDockerEventHandler);
@@ -138,37 +134,35 @@ class Container extends EventEmitter {
     }
 
     /**
-     * @param {string} content 
+     * @param {string} content
      */
     async _send(content) {
-
         const container = this.servers.docker.getContainer(this.name);
         const stream = await container.attach({ stream: true, stdin: true });
 
         await new Promise(async (resolve, reject) => {
             stream.write(content + "\n", (error) => {
                 if (error) reject(error);
-                else stream.end((error) => {
-                    if (error) reject(error);
-                    else resolve();
-                });
+                else
+                    stream.end((error) => {
+                        if (error) reject(error);
+                        else resolve();
+                    });
             });
         });
     }
 
     async kill() {
-
         if (this.state === "stopped") throw new Error("Already stopped");
 
         await this.servers.docker.getContainer(this.name).kill();
     }
 
     /**
-     * @param {string} uuid 
-     * @param {string} username 
+     * @param {string} uuid
+     * @param {string} username
      */
     playerJoin(uuid, username) {
-
         if (this.players.some((player) => player.uuid === uuid)) throw new Error("This player is already connected");
         this.players.push({ uuid, username });
 
@@ -177,10 +171,9 @@ class Container extends EventEmitter {
     }
 
     /**
-     * @param {string} uuid 
+     * @param {string} uuid
      */
     playerQuit(uuid) {
-
         const player = this.players.find((player) => player.uuid === uuid);
 
         if (!player) throw new Error("This player is already disconnected");
@@ -191,13 +184,15 @@ class Container extends EventEmitter {
     }
 
     /**
-     * @param {import("raraph84-lib/src/WebSocketClient")} client 
+     * @param {import("raraph84-lib/src/WebSocketClient")} client
      */
     gatewayConnected(client) {
         if (this.gatewayClient) this.gatewayClient.close("Another client is already connected");
         this.gatewayClient = client;
         this.emit("gatewayConnected");
-        this.servers.gateway.clients.filter((client) => client.infos.logged).forEach((client) => client.emitEvent("SERVER_GATEWAY_CONNECTED", { name: this.name }));
+        this.servers.gateway.clients
+            .filter((client) => client.infos.logged)
+            .forEach((client) => client.emitEvent("SERVER_GATEWAY_CONNECTED", { name: this.name }));
         console.log("Container " + this.name + " connected to the gateway.");
     }
 
@@ -206,24 +201,36 @@ class Container extends EventEmitter {
         while (this.players.length > 0) this.playerQuit(this.players[0].uuid);
         this.gatewayClient = null;
         this.emit("gatewayDisconnected");
-        this.servers.gateway.clients.filter((client) => client.infos.logged).forEach((client) => client.emitEvent("SERVER_GATEWAY_DISCONNECTED", { name: this.name }));
+        this.servers.gateway.clients
+            .filter((client) => client.infos.logged)
+            .forEach((client) => client.emitEvent("SERVER_GATEWAY_DISCONNECTED", { name: this.name }));
         console.log("Container " + this.name + " disconnected from the gateway.");
     }
 
     /**
-     * @param {boolean} logged 
+     * @param {boolean} logged
      */
     toApiObj(logged) {
         return logged
-            ? { name: this.name, port: this.port, state: this.state, gatewayConnected: !!this.gatewayClient, players: this.players.length }
-            : { name: this.name, state: this.state, gatewayConnected: !!this.gatewayClient, players: this.players.length };
+            ? {
+                  name: this.name,
+                  port: this.port,
+                  state: this.state,
+                  gatewayConnected: !!this.gatewayClient,
+                  players: this.players.length
+              }
+            : {
+                  name: this.name,
+                  state: this.state,
+                  gatewayConnected: !!this.gatewayClient,
+                  players: this.players.length
+              };
     }
 }
 
 class Proxy extends Container {
-
     /**
-     * @param {import("./Servers")} servers 
+     * @param {import("./Servers")} servers
      */
     constructor(servers) {
         super(servers, "proxy", path.resolve(config.proxyDir), config.proxyPort);
@@ -239,14 +246,12 @@ class Proxy extends Container {
 }
 
 class Server extends Container {
-
     /**
-     * @param {import("./Servers")} servers 
-     * @param {string} name 
-     * @param {number} maxPlayers 
+     * @param {import("./Servers")} servers
+     * @param {string} name
+     * @param {number} maxPlayers
      */
     constructor(servers, name, maxPlayers) {
-
         let port = config.serversStartingPort;
         while (servers.servers.some((server) => server.port === port)) port++;
 
@@ -257,18 +262,15 @@ class Server extends Container {
     }
 
     async _preStart() {
-
         await super._preStart();
 
-        if (existsSync(this.path))
-            await fs.rm(this.path, { recursive: true });
+        if (existsSync(this.path)) await fs.rm(this.path, { recursive: true });
     }
 
     /**
-     * @param {number} memory 
+     * @param {number} memory
      */
     async start(memory) {
-
         await super.start(memory, "spigot-1.12.2.jar");
     }
 
@@ -277,7 +279,6 @@ class Server extends Container {
     }
 
     playerConnecting(uuid) {
-
         this.connectingPlayers.push(uuid);
 
         const listener = (player) => {
@@ -296,7 +297,7 @@ class Server extends Container {
     }
 
     /**
-     * @param {boolean} logged 
+     * @param {boolean} logged
      */
     toApiObj(logged) {
         return { ...super.toApiObj(logged), maxPlayers: this.maxPlayers };
@@ -304,25 +305,27 @@ class Server extends Container {
 }
 
 class Lobby extends Server {
-
     /**
-     * @param {import("./Servers")} servers 
-     * @param {number} id 
+     * @param {import("./Servers")} servers
+     * @param {number} id
      */
     constructor(servers, id) {
-
         super(servers, "lobby" + id, config.lobbyMaxPlayers);
 
         this.id = id;
     }
 
     async _preStart() {
-
         await super._preStart();
 
         await fs.cp(path.resolve(config.lobbyServerTemplateDir), this.path, { recursive: true });
 
-        const configFile = path.join(this.path, "plugins", "Infinite-Minecraft-Players-POC-Server-Plugin", "config.yml");
+        const configFile = path.join(
+            this.path,
+            "plugins",
+            "Infinite-Minecraft-Players-POC-Server-Plugin",
+            "config.yml"
+        );
         await fs.writeFile(configFile, (await fs.readFile(configFile, "utf8")).replace("SERVERNAME", this.name));
     }
 
@@ -332,13 +335,11 @@ class Lobby extends Server {
 }
 
 class Game extends Server {
-
     /**
-     * @param {import("./Servers")} servers 
-     * @param {number} id 
+     * @param {import("./Servers")} servers
+     * @param {number} id
      */
     constructor(servers, id) {
-
         super(servers, "game" + id, config.gamePlayers);
 
         this.id = id;
@@ -360,12 +361,16 @@ class Game extends Server {
     }
 
     async _preStart() {
-
         await super._preStart();
 
         await fs.cp(path.resolve(config.gameServerTemplateDir), this.path, { recursive: true });
 
-        const configFile = path.join(this.path, "plugins", "Infinite-Minecraft-Players-POC-Server-Plugin", "config.yml");
+        const configFile = path.join(
+            this.path,
+            "plugins",
+            "Infinite-Minecraft-Players-POC-Server-Plugin",
+            "config.yml"
+        );
         await fs.writeFile(configFile, (await fs.readFile(configFile, "utf8")).replace("SERVERNAME", this.name));
     }
 
@@ -379,7 +384,7 @@ class Game extends Server {
     }
 
     /**
-     * @param {boolean} logged 
+     * @param {boolean} logged
      */
     toApiObj(logged) {
         return { ...super.toApiObj(logged), gameStarted: this.gameStarted };
